adding preliminary python build instructions for cmake and improving organization
