+++ momenttensor/c_src/CMakeLists.txt
@@ -17,8 +17,13 @@ SET(CMAKE_BINARY_DIR  ${CMAKE_CURRENT_SOURCE_DIR}/bin)
 SET(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/lib)
 SET(LIBRARY_OUTPUT_PATH ${CMAKE_CURRENT_SOURCE_DIR}/lib)
 IF (COMPEARTH_MAKE_FINTER)
-SET(CMAKE_Fortran_MODULE_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include)
+   SET(CMAKE_Fortran_MODULE_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include)
 ENDIF()
+# For Python
+IF (COMPEARTH_BUILD_PYTHON_MODULE)
+   SET(CMAKE_SWIG_OUTDIR ${CMAKE_SOURCE_DIR}/python)
+ENDIF()
+
 
 IF (COMPEARTH_USE_MKL)
    MESSAGE("Using MKL")
@@ -99,29 +104,28 @@ ADD_TEST(NAME decompositionTest
               WORKING_DIRECTORY ${CMAKE_TESTING_OUTPUT_DIRECTORY}
               COMMAND xdecomp)
 
-INSTALL(TARGETS compearth_shared LIBRARY DESTINATION lib)
-IF (BUILD_PYTHON_MODULE)
-    IF (CMAKE_VERSION VERSION_EQUAL 3.5)
-        FIND_PACKAGE(SWIGOTHER REQUIRED)
-    ELSE (CMAKE_VERSION VERSION_EQUAL 3.5)
-        FIND_PACKAGE(SWIG REQUIRED)
-    ENDIF (CMAKE_VERSION VERSION_EQUAL 3.5)
-    set(CMAKE_SWIG_OUTDIR ${CMAKE_BINARY_DIR}/python)
-    INCLUDE(${SWIG_USE_FILE})
+INSTALL(TARGETS compearth_shared LIBRARY DESTINATION ${LIBRARY_OUTPUT_PATH})
+IF (COMPEARTH_BUILD_PYTHON_MODULE)
+   IF (CMAKE_VERSION VERSION_EQUAL 3.5.1)
+      FIND_PACKAGE(SWIGOTHER REQUIRED)
+   ELSE (CMAKE_VERSION VERSION_EQUAL 3.5.1)
+      FIND_PACKAGE(SWIG REQUIRED)
+      #FIND_PACKAGE(SWIGOTHER REQUIRED)
+   ENDIF (CMAKE_VERSION VERSION_EQUAL 3.5.1)
+   INCLUDE(${SWIG_USE_FILE})
 
-    FIND_PACKAGE(PythonLibs)
-    FIND_PACKAGE(NumPy)
-    INCLUDE_DIRECTORIES(${PYTHON_INCLUDE_DIRS} ${PYTHON_NUMPY_INCLUDE_DIR})
+   FIND_PACKAGE(PythonLibs)
+   FIND_PACKAGE(NumPy)
+   INCLUDE_DIRECTORIES(${PYTHON_INCLUDE_DIRS} ${PYTHON_NUMPY_INCLUDE_DIR})
 
-    IF (CMAKE_VERSION VERSION_LESS 3.0)
-        swig_add_module(compearth python ${CMAKE_SOURCE_DIR}/swig-controller.i)
-    ELSE (CMAKE_VERSION VERSION_LESS 3.0)
-        swig_add_library(compearth LANGUAGE python SOURCES ${CMAKE_SOURCE_DIR}/swig-controller.i)
-    ENDIF (CMAKE_VERSION VERSION_LESS 3.0)
-    swig_link_libraries(compearth compearth_shared ${PYTHON_LIBRARIES})
-    SET_TARGET_PROPERTIES(${SWIG_MODULE_compearth_REAL_NAME} PROPERTIES
-        LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/python)
-    #install(TARGETS ${SWIG_MODULE_compearth_REAL_NAME} LIBRARY DESTINATION
-    #    lib)
-ENDIF (BUILD_PYTHON_MODULE)
+   IF (CMAKE_VERSION VERSION_LESS 3.0)
+      SWIG_ADD_MODULE(compearth python ${CMAKE_SOURCE_DIR}/swig-controller.i)
+   ELSE (CMAKE_VERSION VERSION_LESS 3.0)
+      SWIG_ADD_LIBRARY(compearth LANGUAGE python SOURCES ${CMAKE_SOURCE_DIR}/swig/swig-controller.i)
+   ENDIF (CMAKE_VERSION VERSION_LESS 3.0)
+   SWIG_LINK_LIBRARIES(compearth compearth_shared ${PYTHON_LIBRARIES})
+   SET_TARGET_PROPERTIES(${SWIG_MODULE_compearth_REAL_NAME} PROPERTIES
+                        LIBRARY_OUTPUT_DIRECTORY ${CMAKE_SWIG_OUTDIR})
+   #INSTALL(TARGETS ${SWIG_MODULE_compearth_REAL_NAME} LIBRARY DESTINATION ${CMAKE_SWIG_OUTDIR}))
+ENDIF (COMPEARTH_BUILD_PYTHON_MODULE)
