+++ momenttensor/c_src/swig/numpy.i
