+++ momenttensor/c_src/swig/swig-controller.i
@@ -57,6 +57,7 @@
 %inline %{
 #include "compearth.h"
 #include "compearth_constants.h"
+#include "compearth_enum.h"
 
 int tt2cmt(int ngamma, double *gamma,
            int ndelta, double *delta,
@@ -188,3 +189,4 @@ int h2theta(int nh, double *h, int *ntheta, double **theta)
 
 //#%ignore (compearth_angleMT)
 %include "compearth_constants.h"
+%include "compearth_enum.h"
